# internal/server/server.go
package server

import (
	"context"
	"errors"
	"io/fs"
	"net/http"
	"time"

	"github.com/ryotarai/euphony/internal/agentlog"
	"github.com/ryotarai/euphony/internal/control"
	"github.com/ryotarai/euphony/internal/session"
	webassets "github.com/ryotarai/euphony/web"
)

type Config struct {
	Token              string
	Shell              string
	HookURL            string
	DatabasePath       string
	CodexSessionIndex  string
	CodexSessionsRoot  string
	ClaudeProjectsRoot string
	Assets             fs.FS
}

type Server struct {
	handler   http.Handler
	sessions  *session.Manager
	control   *control.Service
	tickets   *ticketStore
	agentLogs *agentlog.Resolver
}

func New(config Config) (*Server, error) {
	if config.Token == "" {
		return nil, errors.New("EUPHONY_TOKEN is required")
	}

	hooks := session.HookConfig{
		URL:               config.HookURL,
		Token:             config.Token,
		CodexSessionIndex: config.CodexSessionIndex,
	}
	var sessionManager *session.Manager
	var err error
	if config.DatabasePath == "" {
		sessionManager = session.NewManager(config.Shell, hooks)
	} else {
		sessionManager, err = session.NewPersistentManager(config.Shell, hooks, config.DatabasePath)
		if err != nil {
			return nil, err
		}
	}
	tickets := newTicketStore(time.Now)
	controlService, err := control.New(sessionManager)
	if err != nil {
		_ = sessionManager.Close(context.Background())
		return nil, err
	}
	server := &Server{
		sessions:  sessionManager,
		control:   controlService,
		tickets:   tickets,
		agentLogs: agentlog.NewResolver(config.CodexSessionsRoot, config.ClaudeProjectsRoot),
	}

	public := http.NewServeMux()
	public.HandleFunc("GET /api/health", func(w http.ResponseWriter, _ *http.Request) {
		writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
	})
	public.HandleFunc("GET /api/v1/status", v1Status)
	public.HandleFunc("GET /api/v1/schema", v1Schema)
	public.HandleFunc("GET /api/sessions/{id}/terminal", server.terminal)
	public.HandleFunc("GET /api/v1/terminals/{id}/stream", server.v1TerminalStream)

	protected := http.NewServeMux()
	protected.HandleFunc("GET /api/v1/events", server.v1Events)
	protected.HandleFunc("GET /api/v1/terminals", server.v1ListTerminals)
	protected.HandleFunc("POST /api/v1/terminals", server.v1CreateTerminal)
	protected.HandleFunc("GET /api/v1/terminals/{id}", server.v1GetTerminal)
	protected.HandleFunc("DELETE /api/v1/terminals/{id}", server.v1DeleteTerminal)
	protected.HandleFunc("GET /api/v1/terminals/{id}/output", server.v1ReadTerminal)
	protected.HandleFunc("POST /api/v1/terminals/{id}/input", server.v1SendTerminalInput)
	protected.HandleFunc("POST /api/v1/terminals/{id}/run", server.v1RunTerminal)
	protected.HandleFunc("POST /api/v1/terminals/{id}/wait-output", server.v1WaitTerminalOutput)
	protected.HandleFunc("POST /api/v1/terminals/{id}/tickets", server.v1CreateTerminalTicket)
	protected.HandleFunc("GET /api/v1/selection", server.v1GetSelection)
	protected.HandleFunc("PUT /api/v1/selection", server.v1ReplaceSelection)
	protected.HandleFunc("POST /api/v1/selection/actions", server.v1ApplySelection)
	protected.HandleFunc("/api/v1/", func(w http.ResponseWriter, _ *http.Request) {
		writeV1Error(w, http.StatusNotFound, "api_not_found",
			"The API endpoint does not exist.", nil)
	})
	protected.HandleFunc("GET /api/sessions", server.listSessions)
	protected.HandleFunc("POST /api/sessions", server.createSession)
	protected.HandleFunc("DELETE /api/sessions/{id}", server.deleteSession)
	protected.HandleFunc("POST /api/sessions/{id}/acknowledge-attention", server.acknowledgeAttention)
	protected.HandleFunc("POST /api/sessions/{id}/tickets", server.createTicket)
	protected.HandleFunc("GET /api/sessions/{id}/agent-log", server.agentLog)
	protected.HandleFunc("POST /api/hooks/terminal", server.updateTerminalHook)
	protected.HandleFunc("GET /api/settings", server.getSettings)
	protected.HandleFunc("PATCH /api/settings", server.updateSettings)
	protected.HandleFunc("/api/", func(w http.ResponseWriter, _ *http.Request) {
		writeError(w, http.StatusNotFound, "api_not_found", "The API endpoint does not exist.")
	})
	public.Handle("/api/", bearerAuth(config.Token, protected))

	assets := config.Assets
	if assets == nil {
		assets = webassets.Assets
	}
	static, err := newStaticHandler(assets)
	if err != nil {
		static = http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
			w.Header().Set("Content-Type", "text/html; charset=utf-8")
			w.WriteHeader(http.StatusServiceUnavailable)
			_, _ = w.Write(webassets.FallbackHTML)
		})
	}
	public.Handle("/", static)

	server.handler = public
	return server, nil
}

func (s *Server) Handler() http.Handler {
	return s.handler
}

func (s *Server) Close(ctx context.Context) error {
	return s.sessions.Close(ctx)
}